apps/senpy-ai-news-report/pyproject.toml
[tool.poetry]
name = "senpy-ai-chatbot"
version = "0.1.0"
description = ""
authors = ["parabolabam <sorokinvladislav1402@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.13.2"
telethon = "^1.38.1"
apscheduler = "^3.11.0"
requests = "^2.32.3"
bs4 = "^0.0.2"
openai = "^1.60.0"
fastapi = "^0.115.7"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"
aiohttp = "^3.11.11"
debugpy = "^1.8.12"
pydantic = "^2.10.6"
feedparser = "^6.0.11"
pytest = "^8.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
